use a conftest.py for shared blockchain fixtures
