+++ tests/core/full_node/test_full_node_store.py
@@ -20,9 +20,9 @@ from chia.util.block_cache import BlockCache
 from tests.block_tools import get_signage_point, create_block_tools
 from chia.util.hash import std_hash
 from chia.util.ints import uint8, uint32, uint64, uint128
-from tests.core.fixtures import default_1000_blocks, create_blockchain  # noqa: F401
 from tests.setup_nodes import test_constants as test_constants_original
 from tests.util.keyring import TempKeyring
+from tests.util.blockchain import create_blockchain
 
 
 def cleanup_keyring(keyring: TempKeyring):
