+++ tests/core/full_node/test_hint_store.py
@@ -10,7 +10,6 @@ from chia.types.condition_opcodes import ConditionOpcode
 from chia.types.condition_with_args import ConditionWithArgs
 from chia.types.spend_bundle import SpendBundle
 from tests.util.db_connection import DBConnection
-from tests.core.fixtures import empty_blockchain  # noqa: F401
 from tests.wallet_tools import WalletTool
 from tests.setup_nodes import bt
